src/web/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3 